+++ src/Integrations/Airtable/Airtable.php
@@ -11,9 +11,10 @@ declare(strict_types=1);
 namespace EightshiftForms\Integrations\Airtable;
 
 use EightshiftForms\Form\AbstractFormBuilder;
-use EightshiftForms\Integrations\ClientInterface;
 use EightshiftForms\Integrations\MapperInterface;
+use EightshiftFormsVendor\EightshiftFormsUtils\Helpers\UtilsGeneralHelper;
 use EightshiftFormsVendor\EightshiftFormsUtils\Helpers\UtilsHooksHelper;
+use EightshiftFormsVendor\EightshiftLibs\Helpers\Components;
 use EightshiftFormsVendor\EightshiftLibs\Services\ServiceInterface;
 
 /**
@@ -31,16 +32,16 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 	/**
 	 * Instance variable for Airtable data.
 	 *
-	 * @var ClientInterface
+	 * @var AirtableClientInterface
 	 */
 	protected $airtableClient;
 
 	/**
 	 * Create a new instance.
 	 *
-	 * @param ClientInterface $airtableClient Inject Airtable which holds Airtable connect data.
+	 * @param AirtableClientInterface $airtableClient Inject Airtable which holds Airtable connect data.
 	 */
-	public function __construct(ClientInterface $airtableClient)
+	public function __construct(AirtableClientInterface $airtableClient)
 	{
 		$this->airtableClient = $airtableClient;
 	}
@@ -54,6 +55,9 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 	{
 		// Blocks string to value filter name constant.
 		\add_filter(static::FILTER_FORM_FIELDS_NAME, [$this, 'getFormFields'], 10, 3);
+
+		// Recreate dynamic block data for frontend.
+		\add_filter(UtilsHooksHelper::getFilterName(['block', 'dynamic', 'dataOutput']), [$this, 'getDynamicBlockOutput'], 10, 2);
 	}
 
 	/**
@@ -81,7 +85,7 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 			return $output;
 		}
 
-		$fields = $this->getFields($item[$innerId] ?? [], $formId);
+		$fields = $this->getFields($item[$innerId] ?? [], $formId, $item, $itemId);
 
 		if (!$fields) {
 			return $output;
@@ -97,10 +101,12 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 	 *
 	 * @param array<string, mixed> $data Fields.
 	 * @param string $formId Form ID.
+	 * @param array<string, mixed> $items Items.
+	 * @param string $itemId Item ID.
 	 *
 	 * @return array<int, array<string, mixed>>
 	 */
-	private function getFields(array $data, string $formId): array
+	private function getFields(array $data, string $formId, array $items, string $itemId): array
 	{
 		$output = [];
 
@@ -208,6 +214,7 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 					];
 					break;
 				case 'number':
+				case 'currency':
 					$output[] = [
 						'component' => 'input',
 						'inputName' => $name,
@@ -233,11 +240,13 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 					];
 					break;
 				case 'singleSelect':
+				case 'multipleSelects':
 					$output[] = [
 						'component' => 'select',
 						'selectName' => $name,
 						'selectTracking' => $name,
 						'selectFieldLabel' => $label,
+						'selectIsMultiple' => $type === 'multipleSelects',
 						'selectContent' => \array_map(
 							function ($selectOption) {
 								return [
@@ -251,31 +260,8 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 							},
 							$options['choices'] ?? []
 						),
-						'selectDisabledOptions' => $this->prepareDisabledOptions('select'),
-					];
-					break;
-				case 'multipleSelects':
-					$output[] = [
-						'component' => 'checkboxes',
-						'checkboxesName' => $name,
-						'checkboxesTracking' => $name,
-						'checkboxesFieldLabel' => $label,
-						'checkboxesTypeCustom' => 'multiCheckbox',
-						'checkboxesContent' => \array_map(
-							function ($checkbox) {
-								return [
-									'component' => 'checkbox',
-									'checkboxLabel' => $checkbox['name'],
-									'checkboxValue' => $checkbox['id'],
-									'checkboxDisabledOptions' => $this->prepareDisabledOptions('checkbox', [
-										'checkboxValue'
-									], false),
-								];
-							},
-							$options['choices'] ?? []
-						),
-						'checkboxesDisabledOptions' => $this->prepareDisabledOptions('checkboxes', [
-							'checkboxesTypeCustom',
+						'selectDisabledOptions' => $this->prepareDisabledOptions('select', [
+							'selectIsMultiple',
 						]),
 					];
 					break;
@@ -301,6 +287,35 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 						]),
 					];
 					break;
+				case 'multipleRecordLinks':
+					$linkedItemId = $field['options']['linkedTableId'] ?? '';
+					if (!$linkedItemId) {
+						break;
+					}
+
+					$linkedItem = $items[$linkedItemId] ?? [];
+					if (!$linkedItem) {
+						break;
+					}
+
+					$output[] = [
+						'component' => 'dynamic',
+						'dynamicType' => 'select',
+						'dynamicName' => $name,
+						'dynamicTracking' => $name,
+						'dynamicIsDeactivated' => true,
+						'dynamicFieldLabel' => $label,
+						'dynamicTypeCustom' => 'dynamicSelect',
+						// translators: %1$s will be replaced with field label and %2$s with field name.
+						'dynamicCustomLabel' => \sprintf(\__('We will display dynamic data on the frontend for %1$s field - %2$s.', 'eightshift-forms'), $label, $name),
+						'dynamicData' => \wp_json_encode([
+							'baseId' => $itemId,
+							'listId' => $linkedItemId,
+						]),
+						'dynamicDisabledOptions' => $this->prepareDisabledOptions('dynamic'),
+					];
+
+					break;
 			}
 		}
 
@@ -317,5 +332,81 @@ class Airtable extends AbstractFormBuilder implements MapperInterface, ServiceIn
 		}
 
 		return $output;
+	}
+
+	/**
+	 * Recreate dynamic block data for frontend.
+	 *
+	 * @param array<string, mixed> $attributes Attributes.
+	 * @param string $formId Form ID.
+	 *
+	 * @return string
+	 */
+	public function getDynamicBlockOutput(array $attributes, string $formId): string
+	{
+		$formDetails = UtilsGeneralHelper::getFormDetails($formId);
+		$type = $formDetails['type'] ?? '';
+
+		if ($type !== SettingsAirtable::SETTINGS_TYPE_KEY) {
+			return '';
+		}
+
+		$manifest = Components::getComponent('dynamic');
+		$data = Components::checkAttr('dynamicData', $attributes, $manifest);
+
+		if (!$data) {
+			return '';
+		}
+
+		$data = \json_decode($data, true);
+
+		$baseId = $data['baseId'] ?? '';
+		$listId = $data['listId'] ?? '';
+
+		if (!$baseId || !$listId) {
+			return '';
+		}
+
+		$records = $this->airtableClient->getItemDetails($baseId, $listId);
+
+		if (!$records) {
+			return '';
+		}
+
+		$content = '';
+
+		foreach ($records as $record) {
+			$id = $record['id'] ?? '';
+			$title = $record['title'] ?? '';
+
+			if (!$id || !$title) {
+				continue;
+			}
+
+			if (\is_array($title)) {
+				continue;
+			}
+
+			$content .= Components::render(
+				'select-option',
+				[
+					'selectOptionLabel' => $title,
+					'selectOptionValue' => $id,
+				]
+			);
+		}
+
+		return Components::render(
+			'select',
+			Components::props('select', [
+				'selectName' => Components::checkAttr('dynamicName', $attributes, $manifest),
+				'selectTracking' => Components::checkAttr('dynamicTracking', $attributes, $manifest),
+				'selectFieldLabel' => Components::checkAttr('dynamicFieldLabel', $attributes, $manifest),
+				'selectIsMultiple' => Components::checkAttr('dynamicIsMultiple', $attributes, $manifest),
+				'selectIsRequired' => Components::checkAttr('dynamicIsRequired', $attributes, $manifest),
+				'selectTypeCustom' => Components::checkAttr('dynamicTypeCustom', $attributes, $manifest),
+				'selectContent' => $content,
+			])
+		);
 	}
 }